REF: Rename filterset module
