+++ octue/resources/filter_containers.py
